# kallsyms: optimise parsing