In tests with pprof in Inspektor Gadget, I see allocations getting reduced from 34MB to 12MB.

I also added a standalone benchmark.

tl;dr: 12% faster, 42% less allocations in bytes, 48% less allocations

```
$ go test -run=BenchmarkKernelModuleMemory -bench=. ./internal/kallsyms/... -count 10 | tee main.bench
$ go test -run=BenchmarkKernelModuleMemory -bench=. ./internal/kallsyms/... -count 10 | tee patched.bench
$ benchstat main.bench patched.bench
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf/internal/kallsyms
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                     │ main.bench  │            patched.bench            │
                     │   sec/op    │   sec/op     vs base                │
KernelModuleMemory-4   299.9m ± 3%   262.1m ± 6%  -12.59% (p=0.000 n=10)

                     │  main.bench  │            patched.bench             │
                     │     B/op     │     B/op      vs base                │
KernelModuleMemory-4   37.70Mi ± 0%   21.67Mi ± 0%  -42.53% (p=0.000 n=10)

                     │ main.bench  │            patched.bench            │
                     │  allocs/op  │  allocs/op   vs base                │
KernelModuleMemory-4   432.1k ± 0%   221.9k ± 0%  -48.64% (p=0.000 n=10)
```

xref https://github.com/inspektor-gadget/inspektor-gadget/pull/3571#issuecomment-2410532116

cc @burak-ok