@burak-ok 2024-10-15
> so it's a matter of if strings.HasSuffix is faster than bytes.ContainsRune

That is what I meant. Instead of looping through every rune in the string it only needs to check the last one.

> And testing "contains-[" should be more stable than testing "suffix-]", in case Linux adds new fields at the end.

That is what I had feared