@alban 2024-10-16
I rebased my PR to fix the conflict with #1578.

Your idea of custom scanner/iterator looks interesting. Is it something you want to write after this PR gets merged?