@ti-mo 2024-10-17
@alban Closing this in favor of https://github.com/cilium/ebpf/pull/1588. It reduces allocs more or less to the same degree.